Add job and schedule policies. Add worker ID to job. Change permission assignment

+++ app/Http/Controllers/Api/JobsController.php
@@ -6,6 +6,7 @@ use App\Http\Requests;
 use App\Http\Controllers\Controller;
 
 use App\Models\Job;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
@@ -14,10 +15,13 @@ class JobsController extends Controller
     /**
      * Display a listing of the resource.
      *
+     * @param Request $request
      * @return Response
+     * @throws AuthorizationException
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny');
         $jobs = Job::latest()->paginate(25);
 
         return $jobs;
@@ -29,10 +33,11 @@ class JobsController extends Controller
      * @param Request $request
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function store(Request $request)
     {
-        
+        $this->authorize('create');
         $job = Job::create($request->all());
 
         return response()->json($job, 201);
@@ -41,13 +46,15 @@ class JobsController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function show($id)
     {
         $job = Job::findOrFail($id);
+        $this->authorize('view', $job);
 
         return $job;
     }
@@ -56,14 +63,15 @@ class JobsController extends Controller
      * Update the specified resource in storage.
      *
      * @param Request $request
-     * @param  int  $id
+     * @param int $id
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function update(Request $request, $id)
     {
-        
         $job = Job::findOrFail($id);
+        $this->authorize('update', $job);
         $job->update($request->all());
 
         return response()->json($job, 200);
@@ -72,12 +80,15 @@ class JobsController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function destroy($id)
     {
+        $job = Job::findOrFail($id);
+        $this->authorize('delete', $job);
         Job::destroy($id);
 
         return response()->json(null, 204);
